requirements.txt
aioboto3==13.0.1
aiobotocore==2.13.0
aiofiles==23.2.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
attrs==23.2.0
autoflake==2.3.1
bcrypt==4.1.3
billiard==4.2.0
black==24.4.2
boto3==1.34.106
botocore==1.34.106
celery==5.4.0
certifi==2024.6.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cache2==0.2.1
fastapi-cli==0.0.4
flake8==7.1.0
flower==2.0.1
frozenlist==1.4.1
GeoAlchemy==0.7.2
GeoAlchemy2==0.15.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.3.7
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
motor==3.4.0
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.9.1
orjson==3.10.3
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==10.3.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pyasn1==0.6.0
pycodestyle==2.12.0
pydantic==2.7.3
pydantic-extra-types==2.8.0
pydantic-settings==2.3.1
pydantic_core==2.18.4
pyflakes==3.2.0
Pygments==2.18.0
pymongo==4.7.3
pyright==1.1.368
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
time-machine==2.14.1
tornado==6.4.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
